--- a/core/trino-main/src/main/java/io/trino/operator/BigintGroupByHash.java
+++ b/core/trino-main/src/main/java/io/trino/operator/BigintGroupByHash.java
@@ -371,7 +371,8 @@ public class BigintGroupByHash
         return groupId;
     }
 
-    private class AddPageWork
+    @VisibleForTesting
+    class AddPageWork
             implements Work<Void>
     {
         private final Block block;
@@ -412,7 +413,8 @@ public class BigintGroupByHash
         }
     }
 
-    private class AddDictionaryPageWork
+    @VisibleForTesting
+    class AddDictionaryPageWork
             implements Work<Void>
     {
         private final Block dictionary;
@@ -456,7 +458,8 @@ public class BigintGroupByHash
         }
     }
 
-    private class AddRunLengthEncodedPageWork
+    @VisibleForTesting
+    class AddRunLengthEncodedPageWork
             implements Work<Void>
     {
         private final RunLengthEncodedBlock block;
@@ -497,7 +500,8 @@ public class BigintGroupByHash
         }
     }
 
-    private class GetGroupIdsWork
+    @VisibleForTesting
+    class GetGroupIdsWork
             implements Work<GroupByIdBlock>
     {
         private final BlockBuilder blockBuilder;
@@ -546,7 +550,8 @@ public class BigintGroupByHash
         }
     }
 
-    private class GetDictionaryGroupIdsWork
+    @VisibleForTesting
+    class GetDictionaryGroupIdsWork
             implements Work<GroupByIdBlock>
     {
         private final BlockBuilder blockBuilder;
@@ -600,7 +605,8 @@ public class BigintGroupByHash
         }
     }
 
-    private class GetRunLengthEncodedGroupIdsWork
+    @VisibleForTesting
+    class GetRunLengthEncodedGroupIdsWork
             implements Work<GroupByIdBlock>
     {
         private final RunLengthEncodedBlock block;