Few neat tricks to speed up aggregation a bit