@sopel39 2022-02-07
@skrzypo987 could you share perf results?