@skrzypo987 2022-02-08
> @skrzypo987 could you share perf results?

I messed up some things so there is a regression. Will get back with the results once I fix it.