@skrzypo987 2022-02-08
before:
```
BenchmarkGroupByHashOnTpch.groupBy                  false        0  avgt   20    9,134 ±  0,555  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        1  avgt   20   66,969 ±  2,017  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false      2_0  avgt   20   23,235 ±  0,689  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false      2_1  avgt   20   40,110 ±  0,420  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        3  avgt   20   55,257 ±  1,738  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false      4_0  avgt   20   48,726 ±  1,106  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false      4_1  avgt   20   81,607 ±  0,805  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        5  avgt   20   53,240 ±  1,124  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        7  avgt   20  123,469 ±  1,859  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        8  avgt   20    8,932 ±  0,136  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        9  avgt   20   60,057 ±  1,000  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     11_0  avgt   20   22,098 ±  0,386  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     11_1  avgt   20   12,275 ±  0,140  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       12  avgt   20   47,678 ±  1,636  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     13_0  avgt   20   27,317 ±  0,579  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     13_1  avgt   20   37,461 ±  0,424  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       15  avgt   20   22,595 ±  0,226  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     16_0  avgt   20  533,342 ±  4,900  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     16_1  avgt   20  257,857 ±  4,977  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       17  avgt   20   25,246 ±  0,836  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       18  avgt   20   82,138 ±  1,520  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       20  avgt   20   35,989 ±  1,407  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       21  avgt   20  385,947 ± 10,523  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       22  avgt   20   24,116 ±  0,744  ns/op
```
after
```
BenchmarkGroupByHashOnTpch.groupBy                  false        0  avgt   20    6,756 ±  0,049  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        1  avgt   20    2,521 ±  0,034  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false      2_0  avgt   20   19,771 ±  0,672  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false      2_1  avgt   20   39,528 ±  0,643  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        3  avgt   20   47,519 ±  1,084  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false      4_0  avgt   20   48,953 ±  0,771  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false      4_1  avgt   20   69,582 ±  1,493  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        5  avgt   20   51,688 ±  1,756  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        7  avgt   20  127,517 ±  0,964  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        8  avgt   20    6,692 ±  0,035  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false        9  avgt   20   65,012 ±  1,214  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     11_0  avgt   20   17,419 ±  0,184  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     11_1  avgt   20   12,352 ±  0,160  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       12  avgt   20   48,045 ±  0,949  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     13_0  avgt   20   22,845 ±  0,593  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     13_1  avgt   20   35,829 ±  0,841  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       15  avgt   20   19,283 ±  0,421  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     16_0  avgt   20  487,442 ±  5,271  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false     16_1  avgt   20  212,259 ±  4,008  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       17  avgt   20   22,631 ±  0,952  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       18  avgt   20   70,313 ±  1,561  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       20  avgt   20   23,919 ±  1,027  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       21  avgt   20  362,150 ± 10,634  ns/op
BenchmarkGroupByHashOnTpch.groupBy                  false       22  avgt   20   21,731 ±  0,560  ns/op
```