@skrzypo987 2022-02-09
I got rid of the biggest commit as it showed some regressions that were difficult to spot. I'll return to it in different PR.
The only benchmark change with the current state of the PR is tpch/q01 going from ~50ns/row to ~2.5ns/row