@skrzypo987 2022-02-21
Got rid of the commit introducting tpch aggregation benchmark as its newer version exists in https://github.com/trinodb/trino/pull/11031