@sopel39 2022-02-23
Is `BenchmarkGroupByHashOnTpch` part of OS? These benchmark results should be part of commit message